+++ indispensable/src/jvmTest/kotlin/at/asitplus/signum/indispensable/pki/Pkcs10CertificationRequestJvmTest.kt
@@ -103,7 +103,7 @@ class Pkcs10CertificationRequestJvmTest : FreeSpec({
         val signatureAlgorithm = X509SignatureAlgorithm.ES256
         val contentSigner: ContentSigner = signatureAlgorithm.getContentSigner(keyPair.private)
         val spki = SubjectPublicKeyInfo.getInstance(keyPair.public.encoded)
-        val keyUsage = KeyUsage(KeyUsage.digitalSignature)
+        val keyUsage = KeyUsage(X509KeyUsage.digitalSignature)
         val extendedKeyUsage = ExtendedKeyUsage(KeyPurposeId.anyExtendedKeyUsage)
         val bcCsr = PKCS10CertificationRequestBuilder(X500Name("CN=$commonName"), spki)
             .addAttribute(Extension.keyUsage, keyUsage)
@@ -163,7 +163,7 @@ class Pkcs10CertificationRequestJvmTest : FreeSpec({
         val signatureAlgorithm = X509SignatureAlgorithm.ES256
         val contentSigner: ContentSigner = signatureAlgorithm.getContentSigner(keyPair.private)
         val spki = SubjectPublicKeyInfo.getInstance(keyPair.public.encoded)
-        val keyUsage = KeyUsage(KeyUsage.digitalSignature)
+        val keyUsage = KeyUsage(X509KeyUsage.digitalSignature)
         val extendedKeyUsage = ExtendedKeyUsage(KeyPurposeId.anyExtendedKeyUsage)
 
         val extGen = ExtensionsGenerator()
@@ -397,7 +397,7 @@ class Pkcs10CertificationRequestJvmTest : FreeSpec({
             Pkcs10CertificationRequestAttribute
         */
 
-        val keyUsage = KeyUsage(KeyUsage.digitalSignature)
+        val keyUsage = KeyUsage(X509KeyUsage.digitalSignature)
         val extendedKeyUsage = ExtendedKeyUsage(KeyPurposeId.anyExtendedKeyUsage)
 
         val attr1 =